Clean up deep copy for history (#823)
* test: add deep copy tests for StreamParameterization and SimulationState

* feat: add deepCopy to StreamParameterization and SimulationState

* test: add getAtPrior tests for SimulationState linked list

* feat: add getAtPrior and priorState linked list to SimulationState

* style: add JavaDoc and fix checkstyle for getAtPrior

* feat: add grammar rules for get stream N years ago syntax

Add AGO_ token and 4 new grammar rules to support the syntax:
- get streamName N years ago
- get streamName N years ago as units
- get streamName N years ago of "substance"
- get streamName N years ago of "substance" as units

Also add grammar parser tests for all 4 syntax variants.

* test: add GetStreamOperation tests for yearsPast parameter

* feat: add yearsPast field and constructors to GetStreamOperation

* feat: add getStream with yearsPast to Engine interface

* feat: implement getStream with yearsPast in SingleThreadEngine

* fix: correct testExecuteWithYearsPastIndirect to set value on target substance

* style: add JavaDoc and fix checkstyle for yearsPast implementation

* feat: add integration tests and example QTA files for get stream N years ago

* docs: update documentation for get stream N years ago syntax

* refactor: address get-at PR review feedback

- Simplify deepCopy: share immutable EngineNumber refs, add private
  constructor to PriorEquipmentBases instead of direct field mutation
- Remove dead while loop in SimulationState.getStreamDirect (yearsBack
  was always 1) and restructure getAtPrior as if/else if/else if/else
- Drop unnecessary EngineNumber re-wrapping in StreamParameterization
  and SimulationState deepCopy; remove stray whitespace in SingleThreadEngine
- Switch example/docs/tests from 'get priorEquipment N years ago' to
  'get sales N years ago' for correct per-cohort semantics, fixing the
  off-by-one that double-retired the year-1 cohort (now 800/200/0)
- Delete redundant GetStreamPriorGrammarTest (covered by live test)
- Add lead-in sentences and merge paragraphs in tutorial_15


* fix: return typed zero for years-ago queries before sim start

getStream(..., yearsPast) previously returned 0 "" when a years-ago
query predated the simulation start, which broke downstream unit
conversion/arithmetic (e.g. `as units`). Now returns 0 in the requested
conversion units when present, else the stream's native units, using
the existing EngineConstants.getBaseUnits lookup so the mapping cannot
drift from what SimulationState already uses internally.

* refactor: split PriorEquipmentBases into interface + Mutable/Frozen impls

Replaces deepCopy() with freeze(), which returns a FrozenPriorEquipmentBases
snapshot backed entirely by final fields and throwing
UnsupportedOperationException from every mutator, instead of a fully
mutable copy. StreamParameterization now constructs
MutablePriorEquipmentBases and freezes its nested priorEquipmentBases
directly rather than deep-copying it.

* refactor: split StreamParameterization into interface + Mutable/Frozen impls

Replaces deepCopy() with freeze(), which returns a
FrozenStreamParameterization backed entirely by final fields (copied
containers, shared immutable EngineNumber values, recursively frozen
priorEquipmentBases) and throwing UnsupportedOperationException from
every mutator. SimulationState now constructs MutableStreamParameterization
and freezes each substance's parameterization directly during its own
deepCopy rather than deep-copying it via setters.

Updates SimulationStateTest's deepCopy test to reflect that per-substance
parameterization on a copy is now immutable, matching the safety guarantee
this refactor is building toward.

* refactor: split SimulationState into interface + Mutable/Frozen impls

Replaces deepCopy() with freeze(), which returns a FrozenSimulationState
snapshot backed entirely by final fields (copied streams/substances
maps, each substance's StreamParameterization frozen recursively,
priorState captured by reference) and throwing
UnsupportedOperationException from every mutator.

The non-trivial, cross-instance read logic (stream-name dispatch,
sales/virgin/recycle aggregation, the one-year-back lookback, key
building, substance lookup) is extracted into a shared
SimulationStateSupport utility so Mutable and Frozen cannot drift out
of sync; the interface-breaking direct field access a prior state's
`.streams` map used is replaced with public-API calls
(isKnownStream/getStream) so the lookback works across either
implementation.

incrementYear() now freezes the current state to build the prior-year
snapshot instead of deep-copying and manually re-linking priorState -
freeze() already captures it. Adds a regression test asserting that a
setter on a state returned by getAtPrior throws
UnsupportedOperationException, guarding the linked-list safety
property this refactor exists for.

* Address PR #823 review comments

Put String.join's closing paren on its own line, replace duplicated
JavaDoc in MutableSimulationState with {@inheritDoc}, and apply the
preferred multi-line assert* formatting throughout the PR.

---------
